+++ springql-core/src/expression.rs
@@ -232,7 +232,7 @@ impl ValueExprPh2 {
                 SqlValue::NotNull(NnSqlValue::Timestamp(ts)),
                 SqlValue::NotNull(NnSqlValue::Duration(resolution)),
             ) => {
-                let ts_floor = ts.floor(resolution.to_duration());
+                let ts_floor = ts.floor(resolution.to_duration())?;
                 Ok(SqlValue::NotNull(NnSqlValue::Timestamp(ts_floor)))
             }
             _ => Err(SpringError::Sql(anyhow!(