+++ springql-core/src/stream_engine/autonomous_executor/task/pump_task/pump_subtask/query_subtask/group_aggregate_window_subtask.rs
@@ -32,6 +32,7 @@ impl GroupAggregateWindowSubtask {
             .lock()
             .expect("another thread accessing to window gets poisoned")
             .dispatch(expr_resolver, tuple, ())
+            .expect("dispatch failed")
     }
 
     pub fn get_window_mut(&self) -> MutexGuard<AggrWindow> {