+++ CHANGELOG.md
@@ -10,7 +10,11 @@ All other sections are for end-users.
 <!-- markdownlint-disable MD024 -->
 ## [Unreleased]
 
-### For developpers
+### Added
+
+- `BLOB` type ([#187](https://github.com/SpringQL/SpringQL/pull/187))
+
+### For developers
 
 - Add deny lint option for `rustdoc::broken_intra_doc_links` ([#185](https://github.com/SpringQL/SpringQL/pull/185))
 - Refactor : Hide detail module structure ([#177](https://github.com/SpringQL/SpringQL/pull/177))