+++ CHANGELOG.md
@@ -17,6 +17,7 @@ All other sections are for end-users.
     - subsecond part must 9 digits
   - relates security advisory [RUSTSEC-2020-0071](https://rustsec.org/advisories/RUSTSEC-2020-0071)
     - [Tracking issue](https://github.com/SpringQL/SpringQL/issues/173)
+- Bump up MSRV from 1.56.0 to 1.56.1 ([#199](https://github.com/SpringQL/SpringQL/pull/199))
 
 ### Added